Create the function to write the SNFs
Create the function to write the SNFs  to avoid the repeated code in server.js
